requirements.txt
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
boltons==21.0.0
certifi==2022.6.15
charset-normalizer==2.1.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.6
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
face==20.1.1
Faker==13.15.0
glom==22.1.0
greenlet==1.1.2
idna==3.3
MarkupPy==1.14
numpy==1.23.0
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.4.3
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
tablib==3.2.1
termcolor==1.1.0
tzdata==2022.1
urllib3==1.26.9
xlrd==2.0.1
xlwt==1.3.0
